Moved Tenant to trait ilo of Model
