+++ composer.json
@@ -23,7 +23,7 @@
         "illuminate/auth": "~5.3.0|~5.4.0|~5.5.0",
         "illuminate/container": "~5.3.0|~5.4.0|~5.5.0",
         "illuminate/contracts": "~5.3.0|~5.4.0|~5.5.0",
-        "illuminate/database": "~5.4.0|~5.5.0"
+        "illuminate/database": "~5.4.17|~5.5.0"
     },
     "require-dev": {
         "monolog/monolog": "^1.22",
