+++ tests/TenantTest.php
@@ -48,6 +48,6 @@ class TenantTest extends TestCase
     public function it_can_return_a_tenant_object()
     {
         $tenant = app(Tenant::class)->findById(1);
-        $this->assertInstanceOf('\Spatie\Permission\Models\Tenant', $tenant);
+        $this->assertInstanceOf('\Spatie\Permission\Contracts\Tenant', $tenant);
     }
 }
